Move other classes declaration/definition out of eventpipe.*
